Refactor Mapping view with react hooks and clearer components
