+++ client/src/routes.tsx
@@ -3,7 +3,7 @@ import { Route } from "react-router";
 import { BrowserRouter, Switch } from "react-router-dom";
 
 import Authentication from "./views/authentication";
-import Mapping from "./views/mapping";
+import Mapping from "./views/Mapping";
 import MimicView from "./views/mimic";
 import NewSource from "./views/newSource";
 import Sources from "./views/sources";
