+++ client/src/views/Mapping/components/FhirMappingPanel/components/AddResource/resourceNames.tsx
@@ -1,7 +1,3 @@
-// Import custom types
-import { ISimpleAction } from "../../types";
-import { IMappingProps } from "./index";
-
 export const availableResourceNames = [
   {
     type: "Clinical",
@@ -159,83 +155,3 @@ export const availableResourceNames = [
   { type: "Conformance", subtype: "Content", name: "StructureDefinition" },
   { type: "Conformance", subtype: "Content", name: "DataElement" }
 ];
-
-const initialState: IMappingProps = {
-  createdProfiles: 0,
-  createdResources: 0,
-  expandedAttributesIdList: [] as string[],
-  selectedAddResource: {
-    type: null,
-    subtype: null,
-    name: null
-  }
-};
-
-const reducer = (
-  state = initialState,
-  action: ISimpleAction
-): IMappingProps => {
-  switch (action.type) {
-    case "UPDATE_ADD_RESOURCE":
-      return {
-        ...state,
-        selectedAddResource: {
-          type: action.payload.type,
-          subtype: action.payload.subtype,
-          name: action.payload.name
-        }
-      };
-
-    case "ADD_RESOURCE":
-      return {
-        ...state,
-        createdResources: state.createdResources + 1,
-        selectedAddResource: {
-          type: null,
-          subtype: null,
-          name: null
-        }
-      };
-
-    case "DELETE_RESOURCE":
-      console.warn(action.type, action.payload);
-      return {
-        ...state,
-        createdResources: state.createdResources - 1
-      };
-
-    case "ADD_PROFILE":
-      return {
-        ...state,
-        createdProfiles: state.createdProfiles + 1
-      };
-
-    case "DELETE_PROFILE":
-      return {
-        ...state,
-        createdProfiles: state.createdProfiles - 1
-      };
-
-    case "NODE_COLLAPSE":
-      return {
-        ...state,
-        expandedAttributesIdList: state.expandedAttributesIdList.filter(
-          (item: any) => item !== action.payload.nodeData.id
-        )
-      };
-
-    case "NODE_EXPAND":
-      return {
-        ...state,
-        expandedAttributesIdList: [
-          ...state.expandedAttributesIdList,
-          action.payload.nodeData.id
-        ]
-      };
-
-    default:
-      return state;
-  }
-};
-
-export default reducer;
