+++ client/src/views/Mapping/components/SQLRequestParser/index.tsx
