+++ client/src/views/Mapping/components/joinColumns.tsx
@@ -1,5 +1,4 @@
 import * as React from "react";
-import { Mutation, Query, Subscription } from "react-apollo";
 
 import { ISelectedSource } from "../../../types";
