+++ client/src/views/Mapping/components/tabs/SQLRequestParserTab/index.tsx
