+++ client/src/views/Mapping/style.less
